moved sorting logic into a more appropriate home
