+++ app/shared/actions/producers.js
@@ -1,4 +1,5 @@
 import * as types from './types';
+import sortBy from 'lodash/sortBy';
 
 import eos from './helpers/eos';
 
@@ -22,10 +23,22 @@ export function getProducers() {
       scope: 'eosio',
       table: 'producers',
       limit: 1000
-    }).then((results) => dispatch({
-      type: types.GET_PRODUCERS_SUCCESS,
-      payload: { results }
-    })).catch((err) => dispatch({
+    }).then((results) => {
+      const relevant = results.rows.map((producer) => Object.assign({}, {
+        key: `${producer.owner}-${producer.last_produced_block_time}`,
+        last_produced_block_time: producer.last_produced_block_time,
+        owner: producer.owner,
+        url: producer.url,
+        votes: parseInt(producer.total_votes, 10)
+      }));
+      const list = sortBy(relevant, 'votes').reverse();
+      return dispatch({
+        type: types.GET_PRODUCERS_SUCCESS,
+        payload: {
+          list
+        }
+      });
+    }).catch((err) => dispatch({
       type: types.GET_PRODUCERS_FAILURE,
       payload: { err },
     }));
