+++ app/shared/reducers/producers.js
@@ -26,13 +26,7 @@ export default function producers(state = initialState, action) {
     case types.GET_PRODUCERS_SUCCESS: {
       return Object.assign({}, state, {
         __updated: Date.now(),
-        list: action.payload.results.rows.map((producer) => Object.assign({}, {
-          key: `${producer.owner}-${producer.last_produced_block_time}`,
-          last_produced_block_time: producer.last_produced_block_time,
-          owner: producer.owner,
-          url: producer.url,
-          votes: parseInt(producer.total_votes, 10)
-        }))
+        list: action.payload.list
       });
     }
     case types.GET_ACCOUNT_SUCCESS: {
